Backend/requirements.txt
Django~=4.2.7
Django~=4.2.7
django-seed~=0.3.1
psycopg2-binary~=2.9.9
django-debug-toolbar~=4.2.0
django-bootstrap-v5~=1.0.11
django-bootstrap-icons~=0.8.6
django-jquery~=3.1.0
django-bootstrap-datepicker-plus~=5.0.4
django-staticfiles-moment~=2.0.0
django-crispy-forms~=2.0
crispy-bootstrap5~=2023.10
django-environ~=0.11.2
djangorestframework~=3.14.0
django-oauth-toolkit~=2.3.0
django-cors-headers~=4.3.1